Former-commit-id: ab85b19730e61629d3def4487dfa6fed93845e4e [formerly 6e8face49666fb8c3a0692b0bb16619bf7c9ab94]
Former-commit-id: 0b146c1556b2e61b9fa1b9980287dc35bb83d887

+++ app/Http/Controllers/GlobalChargesController.php
@@ -610,7 +610,10 @@ class GlobalChargesController extends Controller
     {
 
         $global = GlobalCharge::find($id);
-        $global->delete();
+        if(!empty($global)){
+            $global->delete();
+        }
+        
     }
     public function editGlobalChar($id)
     {
